+++ src/OpenRiaServices.Tools/Framework/EntityProxyGenerator.cs
@@ -1051,6 +1051,7 @@ namespace OpenRiaServices.Tools
             string[] targetKeyMembers = TypeDescriptor.GetProperties(associationProperty.PropertyType)
                 .Cast<PropertyDescriptor>()
                 .Where(p => p.Attributes[typeof(KeyAttribute)] != null)
+                .OrderBy(p => p.Name, StringComparer.Ordinal)
                 .Select(p => p.Name)
                 .ToArray();