+++ src/Test/Desktop/OpenRiaServices.Common.DomainServices.Test/Baselines/Default/LTS/Northwind_LTS.g.cs
@@ -631,7 +631,7 @@ namespace DataTests.Northwind.LTS
     public sealed partial class Order : Entity
     {
         
-        private EntityRef<Customer> _customer;
+        private EntityRefByKey<Customer> _customer;
         
         private string _customerID;
         
@@ -724,7 +724,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._customer == null))
                 {
-                    this._customer = new EntityRef<Customer>(this, "Customer", this.FilterCustomer);
+                    this._customer = new EntityRefByKey<Customer>(this, "Customer", this.GetCustomerKey);
                 }
                 return this._customer.Entity;
             }
@@ -1176,6 +1176,15 @@ namespace DataTests.Northwind.LTS
             return (entity.CustomerID == this.CustomerID);
         }
         
+        private object GetCustomerKey()
+        {
+            if ((this.CustomerID == null))
+            {
+                return null;
+            }
+            return this.CustomerID;
+        }
+        
         private void AttachOrder_Details(Order_Detail entity)
         {
             entity.Order = this;
@@ -1210,11 +1219,11 @@ namespace DataTests.Northwind.LTS
         
         private float _discount;
         
-        private EntityRef<Order> _order;
+        private EntityRefByKey<Order> _order;
         
         private int _orderID;
         
-        private EntityRef<Product> _product;
+        private EntityRefByKey<Product> _product;
         
         private int _productID;
         
@@ -1287,7 +1296,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._order == null))
                 {
-                    this._order = new EntityRef<Order>(this, "Order", this.FilterOrder);
+                    this._order = new EntityRefByKey<Order>(this, "Order", this.GetOrderKey);
                 }
                 return this._order.Entity;
             }
@@ -1357,7 +1366,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._product == null))
                 {
-                    this._product = new EntityRef<Product>(this, "Product", this.FilterProduct);
+                    this._product = new EntityRefByKey<Product>(this, "Product", this.GetProductKey);
                 }
                 return this._product.Entity;
             }
@@ -1474,9 +1483,19 @@ namespace DataTests.Northwind.LTS
             return (entity.OrderID == this.OrderID);
         }
         
+        private object GetOrderKey()
+        {
+            return this.OrderID;
+        }
+        
         private bool FilterProduct(Product entity)
         {
             return (entity.ProductID == this.ProductID);
+        }
+        
+        private object GetProductKey()
+        {
+            return this.ProductID;
         }
         
         /// <summary>
@@ -1496,7 +1515,7 @@ namespace DataTests.Northwind.LTS
     public sealed partial class Product : Entity
     {
         
-        private EntityRef<Category> _category;
+        private EntityRefByKey<Category> _category;
         
         private Nullable<int> _categoryID;
         
@@ -1583,7 +1602,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._category == null))
                 {
-                    this._category = new EntityRef<Category>(this, "Category", this.FilterCategory);
+                    this._category = new EntityRefByKey<Category>(this, "Category", this.GetCategoryKey);
                 }
                 return this._category.Entity;
             }
@@ -1998,6 +2017,15 @@ namespace DataTests.Northwind.LTS
             return (entity.CategoryID == this.CategoryID);
         }
         
+        private object GetCategoryKey()
+        {
+            if ((this.CategoryID == null))
+            {
+                return null;
+            }
+            return this.CategoryID;
+        }
+        
         private void AttachOrder_Details(Order_Detail entity)
         {
             entity.Product = this;
@@ -2324,7 +2352,7 @@ namespace DataTests.Northwind.LTS
     public sealed partial class Territory : Entity
     {
         
-        private EntityRef<Region> _region;
+        private EntityRefByKey<Region> _region;
         
         private int _regionID;
         
@@ -2367,7 +2395,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._region == null))
                 {
-                    this._region = new EntityRef<Region>(this, "Region", this.FilterRegion);
+                    this._region = new EntityRefByKey<Region>(this, "Region", this.GetRegionKey);
                 }
                 return this._region.Entity;
             }
@@ -2486,6 +2514,11 @@ namespace DataTests.Northwind.LTS
         private bool FilterRegion(Region entity)
         {
             return (entity.RegionID == this.RegionID);
+        }
+        
+        private object GetRegionKey()
+        {
+            return this.RegionID;
         }
         
         /// <summary>