+++ src/Test/Desktop/OpenRiaServices.Common.DomainServices.Test/Baselines/FullTypeNames/Scenarios/MultipleProviderScenarios.g.cs
@@ -621,7 +621,7 @@ namespace DataTests.Northwind.LTS
     public sealed partial class Order : global::OpenRiaServices.Client.Entity
     {
         
-        private global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Customer> _customer;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Customer> _customer;
         
         private string _customerID;
         
@@ -714,7 +714,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._customer == null))
                 {
-                    this._customer = new global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Customer>(this, "Customer", this.FilterCustomer);
+                    this._customer = new global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Customer>(this, "Customer", this.GetCustomerKey);
                 }
                 return this._customer.Entity;
             }
@@ -1166,6 +1166,15 @@ namespace DataTests.Northwind.LTS
             return (entity.CustomerID == this.CustomerID);
         }
         
+        private object GetCustomerKey()
+        {
+            if ((this.CustomerID == null))
+            {
+                return null;
+            }
+            return this.CustomerID;
+        }
+        
         private void AttachOrder_Details(global::DataTests.Northwind.LTS.Order_Detail entity)
         {
             entity.Order = this;
@@ -1200,11 +1209,11 @@ namespace DataTests.Northwind.LTS
         
         private float _discount;
         
-        private global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Order> _order;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Order> _order;
         
         private int _orderID;
         
-        private global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Product> _product;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Product> _product;
         
         private int _productID;
         
@@ -1277,7 +1286,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._order == null))
                 {
-                    this._order = new global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Order>(this, "Order", this.FilterOrder);
+                    this._order = new global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Order>(this, "Order", this.GetOrderKey);
                 }
                 return this._order.Entity;
             }
@@ -1347,7 +1356,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._product == null))
                 {
-                    this._product = new global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Product>(this, "Product", this.FilterProduct);
+                    this._product = new global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Product>(this, "Product", this.GetProductKey);
                 }
                 return this._product.Entity;
             }
@@ -1464,9 +1473,19 @@ namespace DataTests.Northwind.LTS
             return (entity.OrderID == this.OrderID);
         }
         
+        private object GetOrderKey()
+        {
+            return this.OrderID;
+        }
+        
         private bool FilterProduct(global::DataTests.Northwind.LTS.Product entity)
         {
             return (entity.ProductID == this.ProductID);
+        }
+        
+        private object GetProductKey()
+        {
+            return this.ProductID;
         }
         
         /// <summary>
@@ -1486,7 +1505,7 @@ namespace DataTests.Northwind.LTS
     public sealed partial class Product : global::OpenRiaServices.Client.Entity
     {
         
-        private global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Category> _category;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Category> _category;
         
         private global::System.Nullable<int> _categoryID;
         
@@ -1573,7 +1592,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._category == null))
                 {
-                    this._category = new global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Category>(this, "Category", this.FilterCategory);
+                    this._category = new global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Category>(this, "Category", this.GetCategoryKey);
                 }
                 return this._category.Entity;
             }
@@ -1988,6 +2007,15 @@ namespace DataTests.Northwind.LTS
             return (entity.CategoryID == this.CategoryID);
         }
         
+        private object GetCategoryKey()
+        {
+            if ((this.CategoryID == null))
+            {
+                return null;
+            }
+            return this.CategoryID;
+        }
+        
         private void AttachOrder_Details(global::DataTests.Northwind.LTS.Order_Detail entity)
         {
             entity.Product = this;
@@ -2314,7 +2342,7 @@ namespace DataTests.Northwind.LTS
     public sealed partial class Territory : global::OpenRiaServices.Client.Entity
     {
         
-        private global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Region> _region;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Region> _region;
         
         private int _regionID;
         
@@ -2357,7 +2385,7 @@ namespace DataTests.Northwind.LTS
             {
                 if ((this._region == null))
                 {
-                    this._region = new global::OpenRiaServices.Client.EntityRef<global::DataTests.Northwind.LTS.Region>(this, "Region", this.FilterRegion);
+                    this._region = new global::OpenRiaServices.Client.EntityRefByKey<global::DataTests.Northwind.LTS.Region>(this, "Region", this.GetRegionKey);
                 }
                 return this._region.Entity;
             }
@@ -2476,6 +2504,11 @@ namespace DataTests.Northwind.LTS
         private bool FilterRegion(global::DataTests.Northwind.LTS.Region entity)
         {
             return (entity.RegionID == this.RegionID);
+        }
+        
+        private object GetRegionKey()
+        {
+            return this.RegionID;
         }
         
         /// <summary>
@@ -3098,7 +3131,7 @@ namespace NorthwindModel
     public sealed partial class Order : global::OpenRiaServices.Client.Entity
     {
         
-        private global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Customer> _customer;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Customer> _customer;
         
         private string _customerID;
         
@@ -3191,7 +3224,7 @@ namespace NorthwindModel
             {
                 if ((this._customer == null))
                 {
-                    this._customer = new global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Customer>(this, "Customer", this.FilterCustomer);
+                    this._customer = new global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Customer>(this, "Customer", this.GetCustomerKey);
                 }
                 return this._customer.Entity;
             }
@@ -3642,6 +3675,15 @@ namespace NorthwindModel
             return (entity.CustomerID == this.CustomerID);
         }
         
+        private object GetCustomerKey()
+        {
+            if ((this.CustomerID == null))
+            {
+                return null;
+            }
+            return this.CustomerID;
+        }
+        
         private void AttachOrder_Details(global::NorthwindModel.Order_Detail entity)
         {
             entity.Order = this;
@@ -3676,11 +3718,11 @@ namespace NorthwindModel
         
         private float _discount;
         
-        private global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Order> _order;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Order> _order;
         
         private int _orderID;
         
-        private global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Product> _product;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Product> _product;
         
         private int _productID;
         
@@ -3753,7 +3795,7 @@ namespace NorthwindModel
             {
                 if ((this._order == null))
                 {
-                    this._order = new global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Order>(this, "Order", this.FilterOrder);
+                    this._order = new global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Order>(this, "Order", this.GetOrderKey);
                 }
                 return this._order.Entity;
             }
@@ -3823,7 +3865,7 @@ namespace NorthwindModel
             {
                 if ((this._product == null))
                 {
-                    this._product = new global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Product>(this, "Product", this.FilterProduct);
+                    this._product = new global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Product>(this, "Product", this.GetProductKey);
                 }
                 return this._product.Entity;
             }
@@ -3940,9 +3982,19 @@ namespace NorthwindModel
             return (entity.OrderID == this.OrderID);
         }
         
+        private object GetOrderKey()
+        {
+            return this.OrderID;
+        }
+        
         private bool FilterProduct(global::NorthwindModel.Product entity)
         {
             return (entity.ProductID == this.ProductID);
+        }
+        
+        private object GetProductKey()
+        {
+            return this.ProductID;
         }
         
         /// <summary>
@@ -3962,7 +4014,7 @@ namespace NorthwindModel
     public sealed partial class Product : global::OpenRiaServices.Client.Entity
     {
         
-        private global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Category> _category;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Category> _category;
         
         private global::System.Nullable<int> _categoryID;
         
@@ -4049,7 +4101,7 @@ namespace NorthwindModel
             {
                 if ((this._category == null))
                 {
-                    this._category = new global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Category>(this, "Category", this.FilterCategory);
+                    this._category = new global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Category>(this, "Category", this.GetCategoryKey);
                 }
                 return this._category.Entity;
             }
@@ -4463,6 +4515,15 @@ namespace NorthwindModel
             return (entity.CategoryID == this.CategoryID);
         }
         
+        private object GetCategoryKey()
+        {
+            if ((this.CategoryID == null))
+            {
+                return null;
+            }
+            return this.CategoryID;
+        }
+        
         private void AttachOrder_Details(global::NorthwindModel.Order_Detail entity)
         {
             entity.Product = this;
@@ -4786,7 +4847,7 @@ namespace NorthwindModel
     public sealed partial class Territory : global::OpenRiaServices.Client.Entity
     {
         
-        private global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Region> _region;
+        private global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Region> _region;
         
         private int _regionID;
         
@@ -4829,7 +4890,7 @@ namespace NorthwindModel
             {
                 if ((this._region == null))
                 {
-                    this._region = new global::OpenRiaServices.Client.EntityRef<global::NorthwindModel.Region>(this, "Region", this.FilterRegion);
+                    this._region = new global::OpenRiaServices.Client.EntityRefByKey<global::NorthwindModel.Region>(this, "Region", this.GetRegionKey);
                 }
                 return this._region.Entity;
             }
@@ -4944,6 +5005,11 @@ namespace NorthwindModel
         private bool FilterRegion(global::NorthwindModel.Region entity)
         {
             return (entity.RegionID == this.RegionID);
+        }
+        
+        private object GetRegionKey()
+        {
+            return this.RegionID;
         }
         
         /// <summary>